test(selected): selected class에 대한  유닛테스트 작성
selected test의 도메인 테스트, 에러 테스트 분리 및 유닛 테스트 작성
